Added ERC20 wrapper for claim&extClaim
