+++ contracts/utils/WBVRSVoucher.sol
@@ -18,7 +18,7 @@ contract WBVRSVoucher is EIP712 {
     struct NFTVoucher {
         // The id of the token to be redeemed. Must be unique - if another token with this ID already exists, the claim function will revert.
         uint256 tokenId;
-        // In case of ERC20 and ERC1155 the balance could be greater than 1. For ERC721 it must be set to 1.
+        // In case of ERC20 and ERC1155 the balance should be greater than 1. For ERC721 it must be set to 1.
         uint256 balance;
         // The valid nonce value of the NFT creator, fetched through _nonces mapping.
         uint256 nonce;
